requirements.txt
flet
pyrebase4 